Fix: views.py, utils.py
