+++ weneige/users/urls.py
@@ -3,6 +3,6 @@ from .views      import SignUpView, LogInView
 
 urlpatterns = [
     path("/signup", SignUpView.as_view()),
-    path("/login", LogInView.as_view())
+    path("/login", LogInView.as_view()),
 ]
