+++ weneige/weneige/urls.py
@@ -1,5 +1,5 @@
 from django.urls import path, include
 
 urlpatterns = [
-    path("users", include("users.urls"))
+    path("users", include("users.urls")),
 ]
